+++ door-runtime/src/commonMain/kotlin/com/ustadmobile/door/DoorDatabaseSyncRepository.kt
@@ -45,23 +45,7 @@ interface DoorDatabaseSyncRepository: DoorDatabaseRepository {
     suspend fun invalidateAllTables()
 
     /**
-     * Listen for incoming sync changes. This can be used to trigger logic that is required to
-     * update clients (e.g. when a permission change happens granting a client access to an entity
-     * it didn't have access to before).
-     */
-    //fun <T : Any> addSyncListener(entityClass: KClass<T>, listener: SyncListener<T>)
-
-    /**
-     * This is to be called from generated code on the SyncDao's HTTP Endpoint (e.g.
-     * DbNameSyncDao_KtorRoute). It is called after entities are received from an incoming sync. It
-     * will trigger any SyncListeners that were added using addSyncListener
-     *
-     * @param entityClass
-     */
-    //fun <T: Any> handleSyncEntitiesReceived(entityClass: KClass<T>, entities: List<T>)
-
-    /**
      * Do not call this on the main thread: it might run a query
      */
-    val clientId: Int
+    val clientId: Long
 }
